backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
better-auth
psycopg2-binary
pgvector
langchain
langchain-community
langchain-google-genai
tiktoken
asyncpg
google-generativeai
python-multipart

